requirements.txt
amqp==2.2.2
appdirs==1.4.3
billiard==3.5.0.3
cached-property==1.4.2
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.0.2
django-countries==5.2
django-rest-framework==0.1.0
djangorestframework==3.7.7
drf-nested-routers==0.90.2
gunicorn==19.7.1
idna==2.6
isodate==0.6.0
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
Pillow==5.0.0
Pygments==2.2.0
pytz==2018.3
redis==2.10.6
requests==2.18.4
requests-toolbelt==0.8.0
six==1.11.0
spyne==2.12.14
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
zeep==2.5.0
